+++ src/components/sharedComponents/TokenInput/index.tsx
@@ -1,4 +1,4 @@
-import { useEffect, useMemo, useState, type FC } from 'react'
+import { useMemo, type FC } from 'react'
 import styled from 'styled-components'
 
 import { Modal, useModal } from '@faceless-ui/modal'
@@ -28,14 +28,13 @@ import {
   TopRow,
   Wrapper,
 } from '@/src/components/sharedComponents/TokenInput/Components'
+import { UseTokenInputReturnType } from '@/src/components/sharedComponents/TokenInput/useTokenInput'
 import TokenLogo from '@/src/components/sharedComponents/TokenLogo'
 import BaseTokenSelect, {
   Loading,
   type TokenSelectProps,
 } from '@/src/components/sharedComponents/TokenSelect'
-import { useErc20Balance } from '@/src/hooks/useErc20Balance'
 import { type Token } from '@/src/types/token'
-import { isNativeToken } from '@/src/utils/address'
 
 const TokenSelect = styled(BaseTokenSelect)`
   position: relative;
@@ -305,48 +304,6 @@ function TokenAmountField({
       {...(restProps as any)}
     />
   )
-}
-
-export type UseTokenInputReturnType = ReturnType<typeof useTokenInput>
-export function useTokenInput(token?: Token) {
-  const [amount, setAmount] = useState<bigint>(BigInt(0))
-  const [amountError, setAmountError] = useState<string | null>()
-  const [selectedToken, setTokenSelected] = useState<Token | undefined>(token)
-
-  useEffect(() => {
-    setTokenSelected(token)
-  }, [token])
-
-  const { address: userWallet } = useAccount()
-  const { balance, balanceError, isLoadingBalance } = useErc20Balance({
-    address: userWallet ? getAddress(userWallet) : undefined,
-    token: selectedToken,
-  })
-
-  const isNative = selectedToken?.address ? isNativeToken(selectedToken.address) : false
-  const {
-    data: nativeBalance,
-    error: nativeBalanceError,
-    isLoading: isLoadingNativeBalance,
-  } = useBalance({
-    address: userWallet ? getAddress(userWallet) : undefined,
-    chainId: selectedToken?.chainId,
-    query: {
-      enabled: isNative,
-    },
-  })
-
-  return {
-    amount,
-    setAmount,
-    amountError,
-    setAmountError,
-    balance: isNative ? nativeBalance?.value : balance,
-    balanceError: isNative ? nativeBalanceError : balanceError,
-    isLoadingBalance: isNative ? isLoadingNativeBalance : isLoadingBalance,
-    selectedToken,
-    setTokenSelected,
-  }
 }
 
 export default TokenInput
